+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/db/timeseries/questdb/QuestDbFEStatisticsRepository.scala
@@ -3,13 +3,16 @@ package pl.touk.nussknacker.ui.db.timeseries.questdb
 import akka.actor.{ActorSystem, Cancellable}
 import better.files.File
 import cats.effect.{IO, Resource}
+import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
 import io.questdb.cairo.CairoEngine
 import io.questdb.cairo.security.AllowAllSecurityContext
-import io.questdb.griffin.SqlExecutionContextImpl
-import io.questdb.log.LogFactory
-import pl.touk.nussknacker.ui.db.timeseries.FEStatisticsRepository
+import io.questdb.cairo.sql.RecordCursorFactory
+import io.questdb.cairo.wal.WalWriter
+import io.questdb.griffin.{SqlExecutionContext, SqlExecutionContextImpl}
+import pl.touk.nussknacker.ui.db.timeseries.{FEStatisticsRepository, NoOpFEStatisticsRepository}
 import pl.touk.nussknacker.ui.db.timeseries.questdb.QuestDbExtensions.{
+  BuildCairoEngineExtension,
   CairoEngineExtension,
   RecordCursorFactoryExtension
 }
@@ -20,25 +23,18 @@ import pl.touk.nussknacker.ui.db.timeseries.questdb.QuestDbFEStatisticsRepositor
   tableName
 }
 
-import java.nio.charset.StandardCharsets
-import java.nio.file.StandardOpenOption
 import java.time.Clock
 import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
-import java.util.concurrent.{ArrayBlockingQueue, ThreadPoolExecutor, TimeUnit}
+import java.util.concurrent.{ArrayBlockingQueue, ConcurrentHashMap, ThreadPoolExecutor, TimeUnit}
 import scala.concurrent.duration.{Duration, FiniteDuration}
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutorService, Future}
+import scala.jdk.CollectionConverters._
 import scala.util.{Failure, Success, Try}
 
 // TODO list:
-// 2. Compacting?
-// 4. Configurable directory of db (for now it's tmp directory) and limiting db file space on disk.
-// 5. Collecting statistics should be deactivable by configuration. (If db creation fails provide NoOp DB as fallback)
+// 4. Limiting db file space on disk?
 // 6. API should have better types (missing domain layer for FE statistics names).
-// 6. Handling errors while creating the QuestDb.
 // 7. Changing table definition and recreate
-// 9. multi tenant metrics -> add todo that if there are many instances we should change diagrams on grafana
-// 10. flush data in close?
-// todo polish this class in next commits
 private class QuestDbFEStatisticsRepository(private val engine: AtomicReference[CairoEngine], private val clock: Clock)(
     private implicit val ec: ExecutionContextExecutorService
 ) extends FEStatisticsRepository[Future]
@@ -46,32 +42,40 @@ private class QuestDbFEStatisticsRepository(private val engine: AtomicReference[
 
   private val workerCount = 1
 
-  private val sqlContextPool = new ThreadAwareObjectPool(() =>
+  private val sqlContextPool: ThreadAwareObjectPool[SqlExecutionContext] = new ThreadAwareObjectPool(() =>
     new SqlExecutionContextImpl(engine.get(), workerCount).`with`(AllowAllSecurityContext.INSTANCE, null)
   )
 
-  private val recordCursorPool = new ThreadAwareObjectPool(() => engine.get().select(selectQuery, sqlContextPool.get()))
+  private val recordCursorPool: ThreadAwareObjectPool[RecordCursorFactory] =
+    new ThreadAwareObjectPool(() => engine.get().select(selectQuery, sqlContextPool.get()))
 
-  private val walTableWriterPool = new ThreadAwareObjectPool(() =>
+  private val walTableWriterPool: ThreadAwareObjectPool[WalWriter] = new ThreadAwareObjectPool(() =>
     engine.get().getWalWriter(engine.get().getTableTokenIfExists(tableName))
   )
 
   private val flushDataTask      = new AtomicReference(createFlushDataTask())
+  private val taskRecovery       = new TaskRecovery(engine, recover, tableName)
   private lazy val retentionTask = new AtomicReference(createRetentionTask())
-  private val shouldCleanUpData  = new AtomicBoolean(false)
 
-  override def write(statistics: Map[String, Long]): Future[Unit] = run(() => {
-    val statsWalTableWriter = walTableWriterPool.get()
-    statistics.foreach { entry =>
-      val row = statsWalTableWriter.newRow(currentTimeMicros())
-      row.putStr(0, entry._1)
-      row.putLong(1, entry._2)
-      row.append()
+  private val shouldCleanUpData    = new AtomicBoolean(false)
+  private val aggregatedStatistics = new ConcurrentHashMap[String, java.lang.Long]()
+
+  override def write(statistics: Map[String, Long]): Future[Unit] = Future {
+    statistics.foreach { case (key, newValue) =>
+      aggregatedStatistics.compute(
+        key,
+        (_, oldValue) => {
+          if (oldValue == null) {
+            newValue
+          } else {
+            oldValue + newValue
+          }
+        }
+      )
     }
-    statsWalTableWriter.commit()
-  })
+  }
 
-  override def read(): Future[Map[String, Long]] = run(() => {
+  override def read(): Future[Map[String, Long]] = Future {
     recordCursorPool
       .get()
       .fetch(sqlContextPool.get()) { record =>
@@ -80,7 +84,7 @@ private class QuestDbFEStatisticsRepository(private val engine: AtomicReference[
         name -> count
       }
       .toMap
-  })
+  }
 
   private def currentTimeMicros(): Long =
     Math.multiplyExact(clock.instant().toEpochMilli, 1000L)
@@ -97,11 +101,6 @@ private class QuestDbFEStatisticsRepository(private val engine: AtomicReference[
     engine.get().ddl(createTableQuery, sqlContextPool.get())
   }
 
-  private def run[T](dbAction: () => T): Future[T] =
-    Future {
-      engine.get().runWithExceptionHandling(recover, tableName, dbAction)
-    }
-
   private def recover(): Try[Unit] = synchronized {
     Try {
       close()
@@ -113,24 +112,33 @@ private class QuestDbFEStatisticsRepository(private val engine: AtomicReference[
   }
 
   private def flushDataToDisk(): Unit = {
-    flushDataTask.get().run()
+    val statistics          = statisticsSnapshot()
+    val currentTimeInMicros = currentTimeMicros()
+    taskRecovery.runWithRecover(() => flushDataTask.get().runUnsafe(statistics, currentTimeInMicros))
     if (shouldCleanUpData.get()) {
       cleanUpOldData()
     }
   }
 
-  private def cleanUpOldData(): Unit = Try {
+  private def statisticsSnapshot(): Map[String, Long] = {
+    aggregatedStatistics
+      .keySet()
+      .asScala
+      .toList
+      .map(key => key -> Try(aggregatedStatistics.remove(key).longValue()).getOrElse(0L))
+      .toMap
+  }
+
+  private def cleanUpOldData(): Unit = {
     shouldCleanUpData.set(false)
-    retentionTask.get().run(sqlContextPool.get(), clock)
-  }.recover { case ex: Exception =>
-    logger.warn("Exception thrown while cleaning data.", ex)
+    taskRecovery.runWithRecover(() => retentionTask.get().runUnsafe())
   }
 
   private def scheduleRetention(): Unit =
     shouldCleanUpData.set(true)
 
-  private def createFlushDataTask() = new FlushDataTask(engine.get(), tableName)
-  private def createRetentionTask() = new RetentionTask(engine.get(), tableName)
+  private def createFlushDataTask() = new FlushDataTask(engine.get(), walTableWriterPool)
+  private def createRetentionTask() = new RetentionTask(engine.get(), tableName, sqlContextPool, clock)
 }
 
 object QuestDbFEStatisticsRepository extends LazyLogging {
@@ -146,34 +154,78 @@ object QuestDbFEStatisticsRepository extends LazyLogging {
        |    WHERE timestamp_floor('d', ts) = timestamp_floor('d', now())
        | GROUP BY name""".stripMargin
 
-  def create(system: ActorSystem, clock: Clock): Resource[IO, FEStatisticsRepository[Future]] = for {
-    executorService <- createExecutorService()
-    cairoEngine     <- createCairoEngine()
-    repository      <- createRepository(executorService, cairoEngine, clock)
-    // TODO: move task properties to configuration
-    _ <- scheduleTask(system, Duration(2L, TimeUnit.SECONDS), () => repository.flushDataToDisk())
-    _ <- scheduleTask(system, Duration(2L, TimeUnit.SECONDS), () => repository.scheduleRetention())
-  } yield repository
+  def create(system: ActorSystem, clock: Clock, config: Config): Resource[IO, FEStatisticsRepository[Future]] =
+    for {
+      questDbConfig <- Resource.eval(IO(QuestDbConfig.apply(config)))
+      repository <- questDbConfig match {
+        case enabledCfg @ QuestDbConfig.Enabled(_, _, _, _, _) =>
+          createRepositoryResource(system, clock, enabledCfg)
+            .handleErrorWith { t: Throwable =>
+              logger.warn("Creating QuestDb failed", t)
+              createNoOpFEStatisticRepository
+            }
+        case QuestDbConfig.Disabled =>
+          logger.debug("QuestDb is disabled - collecting FE statistics is skipped")
+          createNoOpFEStatisticRepository
+      }
+    } yield repository
 
-  // todo move this ExecutorService properties to a configuration
-  // todo log if task rejected
-  private def createExecutorService(): Resource[IO, ExecutionContextExecutorService] = Resource.make(
-    acquire = for {
-      executorService <- IO(new ThreadPoolExecutor(2, 4, 60L, TimeUnit.SECONDS, new ArrayBlockingQueue[Runnable](8)))
-      ec = ExecutionContext.fromExecutorService(executorService)
-    } yield ec
-  )(release = ec => IO(ec.shutdown()))
+  private def createNoOpFEStatisticRepository: Resource[IO, FEStatisticsRepository[Future]] =
+    Resource.pure[IO, FEStatisticsRepository[Future]](NoOpFEStatisticsRepository)
 
-  private def createCairoEngine(): Resource[IO, AtomicReference[CairoEngine]] = Resource.make(
-    acquire = IO(
-      new AtomicReference(
-        buildCairoEngine(createDirAndConfigureLogging())
-      )
-    )
-  )(release = engine => IO(closeCairoEngine(engine.get())))
+  private def createRepositoryResource(
+      system: ActorSystem,
+      clock: Clock,
+      config: QuestDbConfig.Enabled
+  ): Resource[IO, FEStatisticsRepository[Future]] =
+    for {
+      _               <- Resource.eval(IO(logger.debug(s"QuestDb configuration $config")))
+      executorService <- createExecutorService(config.poolConfig)
+      cairoEngine     <- createCairoEngine(config)
+      repository      <- createRepository(executorService, cairoEngine, clock)
+      _ <- scheduleTask(system, config.flushTaskDelay, config.flushTaskDelay, () => repository.flushDataToDisk())
+      _ <- scheduleTask(system, Duration.Zero, config.retentionTaskDelay, () => repository.scheduleRetention())
+    } yield repository
 
-  private def buildCairoEngine(rootDir: File): CairoEngine =
-    new CairoEngine(new CustomCairoConfiguration(rootDir.canonicalPath))
+  private def createExecutorService(
+      config: QuestDbConfig.QuestDbPoolConfig
+  ): Resource[IO, ExecutionContextExecutorService] =
+    Resource.make(
+      acquire = for {
+        executorService <- IO(
+          new ThreadPoolExecutor(
+            config.corePoolSize,
+            config.maxPoolSize,
+            config.keepAliveTimeInSeconds,
+            TimeUnit.SECONDS,
+            new ArrayBlockingQueue[Runnable](config.queueCapacity)
+          )
+        )
+        ec = ExecutionContext.fromExecutorService(executorService)
+      } yield ec
+    )(release = ec => IO(ec.shutdown()))
+
+  private def createCairoEngine(config: QuestDbConfig.Enabled): Resource[IO, AtomicReference[CairoEngine]] =
+    Resource.make(
+      acquire = IO {
+        val nuDir  = resolveRootDir(config)
+        val engine = buildCairoEngine(nuDir)
+        new AtomicReference(engine)
+      }
+    )(release = engine => IO(closeCairoEngine(engine.get())))
+
+  private def resolveRootDir(config: QuestDbConfig.Enabled): File = {
+    config.directory.map(d => File(d)).getOrElse(File.temp / s"nu/${config.instanceId}")
+  }
+
+  private def buildCairoEngine(rootDir: File): CairoEngine = {
+    logger.debug("Statistics path: {}", rootDir)
+    val canonicalPath = rootDir
+      .createDirIfNotExists()
+      .configureLogging()
+      .canonicalPath
+    new CairoEngine(new CustomCairoConfiguration(canonicalPath))
+  }
 
   private def closeCairoEngine(engine: CairoEngine): Unit =
     Try(engine.close()) match {
@@ -191,54 +243,32 @@ object QuestDbFEStatisticsRepository extends LazyLogging {
       repository <- IO(new QuestDbFEStatisticsRepository(cairoEngine, clock)(ec))
       _          <- IO(repository.createTableIfNotExist())
     } yield repository
-  )(release = repository => IO(repository.close()))
-
-  private def createDirAndConfigureLogging(): File = {
-    val nuDir: File = createRootDirIfNotExists()
-    configureLogging(nuDir)
-    nuDir
-  }
-
-  private def createRootDirIfNotExists(): File = {
-    val nuDir = File.temp.createChild("nu", asDirectory = true)
-    logger.debug("Statistics path: {}", nuDir)
-    nuDir
-  }
-
-  private def configureLogging(rootDir: File): Unit = {
-    rootDir
-      .createChild("conf/log.conf", createParents = true)
-      .writeText(
-        """
-          |writers=stdout
-          |w.stdout.class=io.questdb.log.LogConsoleWriter
-          |w.stdout.level=ERROR
-          |""".stripMargin
-      )(Seq(StandardOpenOption.TRUNCATE_EXISTING), StandardCharsets.UTF_8)
-    LogFactory.configureRootDir(rootDir.canonicalPath)
-  }
+  )(release =
+    repository =>
+      IO {
+        repository.flushDataToDisk()
+        repository.close()
+      }
+  )
 
   private def scheduleTask(
       system: ActorSystem,
+      initialDelay: FiniteDuration,
       interval: FiniteDuration,
       runnable: Runnable
   ): Resource[IO, Cancellable] =
     Resource.make(
       acquire = IO(
         system.scheduler
-          .scheduleWithFixedDelay(interval, interval)(runnable)(system.dispatcher)
+          .scheduleWithFixedDelay(initialDelay, interval)(runnable)(system.dispatcher)
       )
     )(release = task => IO(task.cancel()))
 
   private def recreateCairoEngine(engine: AtomicReference[CairoEngine]): Unit = {
     val cairoEngine = engine.get()
     closeCairoEngine(cairoEngine)
-    val nuDir = File(cairoEngine.getConfiguration.getRoot)
-    if (nuDir.exists) {
-      Try(nuDir.delete())
-    }
-    createDirAndConfigureLogging()
-    engine.set(buildCairoEngine(nuDir))
+    val rootDir = cairoEngine.deleteRootDir()
+    engine.set(buildCairoEngine(rootDir))
   }
 
 }